requirements.txt
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.20.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
aiohttp>=3.9.0
schedule>=1.2.0
rich>=13.0.0
